// ui/mill-grinder-ui/src/api/mill/api.ts
export interface MetadataEntityDto {
    /**
     * 
     * @type {string}
     * @memberof MetadataEntityDto
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof MetadataEntityDto
     */
    'type'?: MetadataEntityDtoTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof MetadataEntityDto
     */
    'schemaName'?: string;
    /**
     * 
     * @type {string}
     * @memberof MetadataEntityDto
     */
    'tableName'?: string;
    /**
     * 
     * @type {string}
     * @memberof MetadataEntityDto
     */
    'attributeName'?: string;
    /**
     * 
     * @type {string}
     * @memberof MetadataEntityDto
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof MetadataEntityDto
     */
    'updatedAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof MetadataEntityDto
     */
    'createdBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof MetadataEntityDto
     */
    'updatedBy'?: string;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof MetadataEntityDto
     */
    'facets'?: { [key: string]: any; };
}

export const MetadataEntityDtoTypeEnum = {
    Catalog: 'CATALOG',
    Schema: 'SCHEMA',
    Table: 'TABLE',
    Attribute: 'ATTRIBUTE',
    Concept: 'CONCEPT'
} as const;

export type MetadataEntityDtoTypeEnum = typeof MetadataEntityDtoTypeEnum[keyof typeof MetadataEntityDtoTypeEnum];